requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.1
altair==5.2.0
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.29.5
azure-identity==1.15.0
backcall==0.2.0
banal==1.0.6
blinker==1.6.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
cryptography==41.0.5
dataclasses-json==0.6.0
dataset==1.6.2
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.13.1
Flask==2.3.3
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-auth==2.23.4
google-search-results==2.4.1
googleapis-common-protos==1.61.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
keyring==24.2.0
langchain==0.0.299
langsmith==0.0.40
lazy-object-proxy==1.9.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
msal==1.24.1
msal-extensions==1.0.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numexpr==2.8.6
numpy==1.26.0
openai==0.27.10
openapi-core==0.18.1
openapi-schema-validator==0.6.2
openapi-spec-validator==0.6.0
opencensus==0.11.3
opencensus-context==0.1.3
opencensus-ext-azure==1.1.11
packaging==23.1
pandas==2.1.2
parse==1.19.1
parso==0.8.3
pathable==0.4.3
pgvector==0.2.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
portalocker==2.8.2
prance==23.6.21.0
prompt-toolkit==3.0.39
promptflow==0.1.0b8
promptflow-tools==0.1.0b10
protobuf==4.25.0
psutil==5.9.5
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.6.3
pydash==6.0.2
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rich==13.7.0
rpds-py==0.10.6
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
semantic-kernel==0.3.10.dev0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sq==1.0a1
SQLAlchemy==1.4.50
stack-data==0.6.2
streamlit==1.29.0
streamlit-chat==0.1.1
strictyaml==1.7.3
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.5
validators==0.22.0
waitress==2.1.2
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.17.0
